+++ cmd/groundcover/api_key.go
@@ -4,8 +4,9 @@ import (
 	"fmt"
 
 	"github.com/getsentry/sentry-go"
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	"groundcover.com/pkg/auth"
+	"groundcover.com/pkg/api"
 )
 
 var apiKeyCmd = &cobra.Command{
@@ -14,9 +15,9 @@ var apiKeyCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		var err error
 
-		var apiKey *auth.ApiKey
-		if apiKey, err = auth.LoadApiKey(); err != nil {
-			return err
+		var apiKey api.ApiKey
+		if err = apiKey.Load(); err != nil {
+			return errors.Wrap(err, "failed to load api key")
 		}
 
 		fmt.Println(apiKey.ApiKey)