+++ CHANGELOG.md
@@ -20,7 +20,11 @@ We use the following categories for changes:
 
 ### Changed
 
+- split authentication functionality to api and auth clients [#sc-4313]
+
 ### Fixed
+
+- validate or refresh auth0 token per api request [#sc-4313]
 
 ### Removed